+++ dao/src/main/java/com/gjj/igden/dao/AccountRowMapper.java
@@ -9,7 +9,7 @@ import java.sql.SQLException;
 public class AccountRowMapper implements RowMapper<Account> {
   public Account mapRow(ResultSet resultSet, int i) throws SQLException {
     Account account = new Account();
-    account.setId(resultSet.getInt("account_id"));
+    account.setId(resultSet.getLong("account_id"));
     account.setAccountName(resultSet.getString("account_name"));
     account.setEmail(resultSet.getString("email"));
     account.setAdditionalInfo(resultSet.getString("additional_info"));
